backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
python-multipart==0.0.6
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.2
torchvision==0.16.2
Pillow==10.1.0
google-generativeai==0.3.1
httpx==0.25.2
